Improved traps
We are now trapping SIGINT, SIGTERM, SIGHUP and SIGQUIT. When trapped, `handle_signals` is called.
We are also trapping SIGEXIT, which calls `handle_exit`, the new exit door of the WABACMachine.

Note that rsync returns 20 when it receives SIGUSR1 or SIGINT. We transform this into a `RSYNC_INTERRUPT` error.

Also removed `cleanupexit` function and replaced it by `error_exit`.